voxelands V0.2-20240715 Date : 15 juillet 2024 Auteur : Pierre Brochard (pierre.brochard.1982@m4x.org) Copyright et droits d'auteurs : Pierre Brochard

+++ src/crypto_base64.c
@@ -186,7 +186,7 @@ size_t base64_ldecode(const char* source, char *target, size_t targetlen)
 		tmplen = s_base64_decode_triple((unsigned char*)quadruple, (unsigned char*)tmpresult);
 
 		/* check if the fit in the result buffer */
-		if (targetlen < tmplen) {
+		if (targetlen < (size_t) tmplen) {
 			free(src);
 			return -1;
 		}
